--- a/core/pfe-core/controllers/floating-dom-controller.ts
+++ b/core/pfe-core/controllers/floating-dom-controller.ts
@@ -1,15 +1,70 @@
-import { computePosition, shift, arrow, offset, autoUpdate } from '@floating-ui/dom';
 import type { Placement } from '@floating-ui/dom';
 import type { ReactiveController, ReactiveElement } from 'lit';
+import type { StyleInfo } from 'lit/directives/style-map.js';
+import type { Options as Offset } from '@floating-ui/core/src/middleware/offset';
+
+
+export { Placement };
+
+import {
+  autoUpdate,
+  computePosition,
+  offset as offsetMiddleware,
+  shift as shiftMiddleware,
+  flip as flipMiddleware,
+} from '@floating-ui/dom';
+
+type Lazy<T> = T|(() => T|null|undefined);
+
+interface FloatingDOMControllerOptions {
+  content: Lazy<HTMLElement>;
+  invoker?: Lazy<HTMLElement>;
+  arrow?: boolean;
+  flip?: boolean;
+  shift?: boolean;
+  padding?: number;
+}
+
+interface ShowOptions {
+  offset?: Offset;
+  placement?: Placement;
+}
+
+export type Anchor = ''|'top'|'left'|'bottom'|'right';
+export type Alignment = 'center'|'start'|'end';
 
 /**
  * Controls floating DOM within a web component, e.g. tooltips and popovers
  */
 export class FloatingDOMController implements ReactiveController {
   #open = false;
-  #initialized = false;
+  #opening = false;
   #cleanup?: () => void;
-  #calcPosition = true;
+  #anchor?: Anchor;
+  #alignment?: Alignment;
+  #styles?: StyleInfo;
+  #placement?: Placement;
+  #options: Required<FloatingDOMControllerOptions>;
+
+  get #invoker() {
+    const { invoker } = this.#options;
+    return typeof invoker === 'function' ? invoker() : invoker;
+  }
+
+  get #content() {
+    const { content } = this.#options;
+    return typeof content === 'function' ? content() : content;
+  }
+
+  /** The crosswise alignment of the invoker on which to display the floating DOM */
+  get alignment() {
+    return this.#alignment ?? 'center';
+  }
+
+  /** The side of the invoker on which to display the floating DOM */
+  get anchor() {
+    return this.#anchor ?? '';
+  }
 
   /**
    * When true, the floating DOM is visible
@@ -18,126 +73,90 @@ export class FloatingDOMController implements ReactiveController {
     return this.#open;
   }
 
-  set open(value: boolean) {
-    this.#open = value;
-
-    this.host.requestUpdate();
+  /** The computed placement of the floating DOM */
+  get placement(): Placement {
+    return this.#placement ?? 'top';
   }
 
-  get initialized() {
-    return this.#initialized;
+  /**
+   * Styles to apply to your element's container
+   *
+   * - `--_floating-content-translate`: translate to apply to floating content.
+   */
+  get styles(): StyleInfo {
+    return this.#styles ?? {};
   }
 
-  set initialized(v: boolean) {
-    this.#initialized = v; this.host.requestUpdate();
-  }
-
-  get cleanup() {
-    return this.#cleanup;
-  }
-
-  set cleanup(v: any) {
-    this.#cleanup = v;
-    this.host.requestUpdate();
-  }
-
-  get calcPosition() {
-    return this.#calcPosition;
-  }
-
-  set calcPosition(v: any) {
-    this.#calcPosition = v;
-    this.host.requestUpdate();
-  }
-
-  setAutoUpdate(invoker: Element, content: HTMLElement) {
-    if (!this.cleanup) {
-      this.cleanup = autoUpdate(
-        invoker,
-        content,
-        () => {
-          if (!this.calcPosition) {
-            this.calcPosition = true;
-          }
-        }
-      );
-    }
-  }
-
-  removeAutoUpdate() {
-    if (this.#cleanup) {
-      this.#cleanup();
-    }
-  }
-
-  constructor(private host: ReactiveElement) {
+  constructor(
+    private host: ReactiveElement,
+    options: FloatingDOMControllerOptions
+  ) {
     host.addController(this);
+    this.#options = options as Required<FloatingDOMControllerOptions>;
+    this.#options.invoker ??= host;
+    this.#options.arrow ??= false;
+    this.#options.flip ??= true;
+    this.#options.shift ??= true;
   }
 
-  hostConnected?(): void;
+  hostDisconnected() {
+    this.#cleanup?.();
+  }
+
+  async #update(placement: Placement = 'top', offset?: Offset) {
+    const { flip, padding, shift } = this.#options;
+
+    const invoker = this.#invoker;
+    const content = this.#content;
+    if (!invoker || !content) {
+      return;
+    }
+    const { x, y, placement: _placement } = await computePosition(invoker, content, {
+      strategy: 'absolute',
+      placement,
+      middleware: [
+        offsetMiddleware(offset),
+        shift && shiftMiddleware({ padding }),
+        flip && flipMiddleware({ padding }),
+      ].filter(Boolean)
+    });
+
+    this.#placement = _placement;
+    [this.#anchor, this.#alignment] = (this.#placement.split('-') ?? []) as [Anchor, Alignment];
+    this.#styles = {
+      '--_floating-content-translate': `${x}px ${y}px`,
+    };
+    this.host.requestUpdate();
+  }
 
   /** Show the floating DOM */
-  show(): void {
-    this.open = true;
+  async show({ offset, placement }: ShowOptions = {}) {
+    const invoker = this.#invoker;
+    const content = this.#content;
+    if (!invoker || !content) {
+      return;
+    }
+    if (!this.#opening) {
+      this.#opening = true;
+      const p = this.#update(placement, offset);
+      this.#cleanup ??= autoUpdate(invoker, content, () =>
+        this.#update(placement, offset));
+      await p;
+      this.#opening = false;
+    }
+    this.#open = true;
+    this.host.requestUpdate();
   }
 
   /** Hide the floating DOM */
-  hide(): void {
-    this.open = false;
-  }
-
-  computeElementPosition(invoker: Element, content: HTMLElement, arrowElement: HTMLElement, providedPlacement: Placement = 'left', offsetNumbers?: number[]) {
-    if (invoker && content) {
-      computePosition(invoker!, content, {
-        strategy: 'absolute',
-        placement: providedPlacement,
-        middleware: [
-          offset(offsetNumbers ? offsetNumbers[0] : 0),
-          shift(),
-          arrow({
-            element: arrowElement
-          }),
-        ]
-      }).then(({ x, y, placement, middlewareData }) => {
-        Object.assign(content!.style, {
-          top: `${y}px`,
-          left: `${x}px`,
-        });
-
-        if (middlewareData.arrow) {
-          const { y: arrowY, x: arrowX } = middlewareData.arrow;
-
-          // @ts-ignore
-          let staticSide = {
-            top: 'bottom',
-            right: 'left',
-            bottom: 'top',
-            left: 'right',
-          }[placement.split('-')[0]];
-
-          if (staticSide === undefined) {
-            staticSide = 'left';
-          }
-
-          Object.assign(arrowElement!.style, {
-            left: arrowX != null ? `${arrowX}px` : '',
-            top: arrowY != null ? `${arrowY}px` : '',
-            right: '',
-            bottom: '',
-            [staticSide]: '-4px',
-          });
-        }
-      });
-      this.calcPosition = false;
+  async hide() {
+    await this.host.updateComplete;
+    while (this.#opening && !this.open) {
+      await new Promise(requestAnimationFrame);
     }
-    this.initialized ||= true;
-  }
-
-  /** Initialize the floating DOM */
-  create(invoker: Element, content: HTMLElement, arrowElement: HTMLElement, providedPlacement: any = 'left', offsetNumbers?: number[]): void {
-    if (invoker && content) {
-      this.setAutoUpdate(invoker, content);
-      this.computeElementPosition(invoker, content, arrowElement, providedPlacement, offsetNumbers);
-    }
+    this.#open = false;
+    this.#cleanup?.();
+    this.host.requestUpdate();
+    await this.host.updateComplete;
   }
 }