--- a/elements/pfe-tooltip/BaseTooltip.scss
+++ b/elements/pfe-tooltip/BaseTooltip.scss
@@ -1,103 +1,23 @@
-[class*=pf-c-], [class*=pf-c-]::before, [class*=pf-c-]::after {
-  padding: 0;
-  margin: 0;
-  background-color: transparent;
-}
-
 #tooltip {
   color: var(--pf-c-tooltip__content--Color,
     var(--pf-global--Color--light-100, #ffffff));
   background-color: var(--pf-c-tooltip__content--BackgroundColor,
     var(--pf-global--BackgroundColor--dark-100, #151515));
   line-height: var(--pf-c-tooltip--line-height, 1.5);
-  opacity: 1;
-  transition: opacity 300ms cubic-bezier(0.54, 1.5, 0.38, 1.11) 0s;
-  position: relative;
+  position: absolute;
   max-width: var(--pf-c-tooltip--MaxWidth, 18.75rem);
   box-shadow: var(--pf-c-tooltip--BoxShadow,
     var(--pf-global--BoxShadow--md,
       0 0.25rem 0.5rem 0rem rgba(3, 3, 3, 0.12),
       0 0 0.25rem 0 rgba(3, 3, 3, 0.06)));
-  z-index: 999;
-}
-
-#tooltip:not(.initialized) { display: none; }
-
-#tooltip[aria-hidden="true"] {
+  z-index: 10000;
   opacity: 0;
   transition: opacity 300ms cubic-bezier(0.54, 1.5, 0.38, 1.11) 0s;
   pointer-events: none;
 }
 
-/**
-  NOTE: If we replace the popper library the styles related to anything data-popper-* will need to be updated
-        These are popper specific data attributes.
-**/
-
-[data-popper-placement='top'] .arrow,
-[data-popper-placement='top-end'] .arrow,
-[data-popper-placement='top-start'] .arrow {
-  bottom: 0;
-  left: 50%;
-  transform:
-    translateX(var(--pf-c-tooltip__arrow--m-top--TranslateX, -50%))
-    translateY(var(--pf-c-tooltip__arrow--m-top--TranslateY, 50%))
-    rotate(var(--pf-c-tooltip__arrow--m-top--Rotate, 45deg));
-}
-
-[data-popper-placement='bottom'] .arrow,
-[data-popper-placement='bottom-end'] .arrow,
-[data-popper-placement='bottom-start'] .arrow {
-  top: 0;
-  left: 50%;
-  transform:
-    translateX(var(--pf-c-tooltip__arrow--m-bottom--TranslateX, -50%))
-    translateY(var(--pf-c-tooltip__arrow--m-bottom--TranslateY, -50%))
-    rotate(var(--pf-c-tooltip__arrow--m-bottom--Rotate, 45deg));
-}
-
-[data-popper-placement='left'] .arrow,
-[data-popper-placement='left-end'] .arrow,
-[data-popper-placement='left-start'] .arrow {
-  top: 50%;
-  right: 0;
-  transform:
-    translateX(var(--pf-c-tooltip__arrow--m-left--TranslateX, 50%))
-    translateY(var(--pf-c-tooltip__arrow--m-left--TranslateY, -50%))
-    rotate(var(--pf-c-tooltip__arrow--m-left--Rotate, 45deg));
-}
-
-[data-popper-placement='right'] .arrow,
-[data-popper-placement='right-end'] .arrow,
-[data-popper-placement='right-start'] .arrow {
-  top: 50%;
-  left: 0;
-  transform:
-    translateX(var(--pf-c-tooltip__arrow--m-right--TranslateX, -50%))
-    translateY(var(--pf-c-tooltip__arrow--m-right--TranslateY, -50%))
-    rotate(var(--pf-c-tooltip__arrow--m-right--Rotate, 45deg));
-}
-
-[data-popper-placement='left-start'] .arrow,
-[data-popper-placement='right-start'] .arrow {
-  top: var(--pf-c-tooltip__arrow--Height, 0.5rem);
-}
-
-[data-popper-placement='left-end'] .arrow,
-[data-popper-placement='right-end'] .arrow {
-  top: auto;
-  bottom: 0;
-}
-
-[data-popper-placement='top-start'] .arrow,
-[data-popper-placement='bottom-start'] .arrow {
-  left: var(--pf-c-tooltip__arrow--Width, 0.5rem);
-}
-
-[data-popper-placement='top-end'] .arrow,
-[data-popper-placement='bottom-end'] .arrow {
-  right: 0;
-  left: auto;
+:host([showing]) #tooltip {
+  opacity: 1;
 }
 
 .content {
@@ -118,18 +38,21 @@
   word-break: break-word;
   background-color: var(--pf-c-tooltip__content--BackgroundColor,
     var(--pf-global--BackgroundColor--dark-100, #151515));
+  // z-index: 10000; 
 }
 
-.arrow {
+#arrow {
   position: absolute;
   width: var(--pf-c-tooltip__arrow--Width, 0.5rem);
   height: var(--pf-c-tooltip__arrow--Height, 0.5rem);
   pointer-events: none;
   background-color: var(--pf-c-tooltip__content--BackgroundColor,
     var(--pf-global--BackgroundColor--dark-100, #151515));
+  transform: rotate(45deg)
 }
 
 #invoker {
   width: fit-content;
   display: inline-block;
 }
+