--- a/elements/pfe-tooltip/BaseTooltip.scss
+++ b/elements/pfe-tooltip/BaseTooltip.scss
@@ -38,7 +38,6 @@
   word-break: break-word;
   background-color: var(--pf-c-tooltip__content--BackgroundColor,
     var(--pf-global--BackgroundColor--dark-100, #151515));
-  // z-index: 10000; 
 }
 
 #arrow {