--- a/tools/pfe-tools/_scripts/dev.js
+++ b/tools/pfe-tools/_scripts/dev.js
@@ -11,8 +11,5 @@ esbuild.build({
   watch: true,
   color: true,
   outdir: cwd,
-  define: {
-    'process.env.NODE_ENV': 'production'
-  },
   entryPoints,
 });