### What has changed and why

Currently tooltip has a dependency on the [popperjs library](https://popper.js.org/) that while it works it has some issues for us, notably
- Performance
  - `@popperjs/core` has some  
- The `process.env.NODE_ENV` issue 
  - Currently when `pfe-tooltip` or `rh-tooltip` are being used we need to make sure the `process.env.NODE_ENV` variable is set to `production` so the @popperjs library can function. 

As a result, we want to make the switch to using `@floating-ui/dom` which will provide all the same functionality as `@popperjs/core` but without the `NODE_ENV` issue and with some small performance enhancements.

![image](https://user-images.githubusercontent.com/9708152/206573200-8e28034f-8e6b-4aab-b62f-5a759c400026.png)

### Performance Tests

| Test        | Current          | New ✨  |
| ------------- |:-------------:| -----:|
| Rendering 10,000 elements      | 0.d5478s | 0.4255s |
| Triggering .show() on 10,000 elements      | 117s (29.434s for 2500, x4)      |   6.744s |
| Triggering .click() back and forth between two tooltips 2,000 times | 1.12s     |    0.37s |

### Testing Instructions

Closes #2258. Ensure this PR addresses accessibility concerns raised there.